feat: add footer
